requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# AI & ML
openai==1.40.0

# Frontend
streamlit==1.28.0

# Image processing
pillow==10.1.0

# Environment variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
httpx>=0.27.0
